# Add a seralization group to reduce the config when serializing