--- a/src/Sulu/Bundle/AdminBundle/Metadata/FormMetadata/FormMetadata.php
+++ b/src/Sulu/Bundle/AdminBundle/Metadata/FormMetadata/FormMetadata.php
@@ -11,6 +11,7 @@
 
 namespace Sulu\Bundle\AdminBundle\Metadata\FormMetadata;
 
+use JMS\Serializer\Annotation\Exclude;
 use JMS\Serializer\Annotation\SerializedName;
 use Sulu\Bundle\AdminBundle\Metadata\AbstractMetadata;
 use Sulu\Bundle\AdminBundle\Metadata\SchemaMetadata\SchemaMetadata;
@@ -19,11 +20,15 @@ class FormMetadata extends AbstractMetadata
 {
     /**
      * @var string
+     *
+     * @Exclude(if="'admin_form_metadata_keys_only' in context.getAttribute('groups')")
      */
     private $name;
 
     /**
      * @var string
+     *
+     * @Exclude(if="'admin_form_metadata_keys_only' in context.getAttribute('groups')")
      */
     private $title;
 
@@ -31,21 +36,28 @@ class FormMetadata extends AbstractMetadata
      * @var ItemMetadata[]
      *
      * @SerializedName("form")
+     * @Exclude(if="'admin_form_metadata_keys_only' in context.getAttribute('groups')")
      */
     private $items;
 
     /**
      * @var SchemaMetadata
+     *
+     * @Exclude(if="'admin_form_metadata_keys_only' in context.getAttribute('groups')")
      */
     private $schema;
 
     /**
      * @var string
+     *
+     * @Exclude(if="'admin_form_metadata_keys_only' in context.getAttribute('groups')")
      */
     private $key;
 
     /**
      * @var TagMetadata[]
+     *
+     * @Exclude(if="'admin_form_metadata_keys_only' in context.getAttribute('groups')")
      */
     protected $tags;